Séparation des responsabilités
